refactor: articleId now is a lon

+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/ArticleService.java
@@ -12,7 +12,7 @@ public interface ArticleService {
 
     Optional<Article> createArticle(String title, String description, Float pricePerDay, List<Long> categories, List<MultipartFile> image, long idOwner);
 
-    Optional<Article> findById(Integer articleId);
+    Optional<Article> findById(long articleId);
 
     Long getMaxPage(String name, Long category, Long user, Long location);
 }
